% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_util.R
\name{baseRPlotting}
\alias{baseRPlotting}
\title{Plot Second Order vs First Order minimal depth of a maximal subtree depth
for base R}
\usage{
baseRPlotting(eval_data)
}
\arguments{
\item{eval_data}{The output of calculateAMDMS.}
}
\value{
plot object
}
\description{
Base r version of plotFirstAndSecondOrderMetric (just an alternative).
}
