% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_traversal_util.R
\name{countSplitsPerVar}
\alias{countSplitsPerVar}
\title{Count Splits Per Variable}
\usage{
countSplitsPerVar(ranger_obj_forest)
}
\arguments{
\item{ranger_obj_forest}{A ranger object from the ranger package, which was created
with param write.forest set to TRUE. In other words, it must have a 
'forest' property.}
}
\value{
A dataframe with one column of counts, and one column of
normalized counts. Rows are labeled by variable names.
}
\description{
This function counts the number of times each variable was used to split a
tree.
}

