% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_traversal_util.R
\name{forestAvgMaxSubtree}
\alias{forestAvgMaxSubtree}
\title{Forest Averaged Maximal Subtree}
\usage{
forestAvgMaxSubtree(ranger_obj)
}
\arguments{
\item{ranger_obj}{a ranger object from the ranger package, which was created
by setting param write.forest to TRUE. In other words, it must have a 
'forest' property.}
}
\value{
a data.frame with two columns: averaged first and second order 
minimal depth of a maximal subtree
}
\description{
Given a result from the Ranger package (write.forest must
be set to TRUE), this function will traverse the trees and calculate the
first and second order average minimal depth of a maximal subtree.
}

