% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_util.R
\name{getAndPlotFirstAndSecondOrderMetric}
\alias{getAndPlotFirstAndSecondOrderMetric}
\title{Calculate the metric, and make Second Ord. vs First Ord. plot}
\usage{
getAndPlotFirstAndSecondOrderMetric(ranger_result)
}
\arguments{
\item{ranger_result}{ranger result with a saved $forest object.
(set write.forest to TRUE)}
}
\value{
a list: element 1 is a data.frame containing subtree depth
data, element 2 is the plot
}
\description{
comprehensive function for calculating minimal depth of a maximal 
subtree, averaged over the forest, and then plotting the result
}
\examples{
require(survival)
library(ranger)
rg.veteran <- ranger(Surv(time, status) ~ ., data = veteran, write.forest =
TRUE)
result <- getAndPlotFirstAndSecondOrderMetric(rg.veteran)
}

