% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_util.R
\name{getAndPlotSOvsFO}
\alias{getAndPlotSOvsFO}
\title{Calculate subtree depth metric, and make Second Ord. vs First Ord. plot}
\usage{
getAndPlotSOvsFO(ranger_result)
}
\arguments{
\item{ranger_result}{ranger result with a saved $forest object.
(set write.forest to TRUE)}
}
\value{
a list: element 1 is a data.frame containing subtree depth
data, element 2 is the plot
}
\description{
comprehensive function for calculating subtree metric, and then
plotting the result
}
\examples{
require(survival)
library(ranger)
rg.veteran <- ranger(Surv(time, status) ~ ., data = veteran, write.forest =
TRUE)
result <- getAndPlotSOvsFO(rg.veteran)
}

