% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_util.R
\name{plotAMDMSvsNumSplits}
\alias{plotAMDMSvsNumSplits}
\title{Plot metric vs number of splits per variable.}
\usage{
plotAMDMSvsNumSplits(eval_data)
}
\arguments{
\item{eval_data}{The output of calculateAMDMS.}
}
\value{
plot object
}
\description{
Plots the metric against the number of splits in the forest. Features with
many splits will be weighed unfairly by the metric.
}

