% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_util.R
\name{plotFirstAndSecondOrderMetric}
\alias{plotFirstAndSecondOrderMetric}
\title{Plot Second Order vs First Order, minimal depth of a maximal subtree,
averaged over the forest (shortened to "metric" for brevity)}
\usage{
plotFirstAndSecondOrderMetric(eval_data)
}
\arguments{
\item{eval_data, }{the output of forestAvgMaxSubtree}
}
\value{
so_vs_fo, a ggplot2 object, which plots second order vs first
order minimal maximal subtree depth
}
\description{
Given eval data from forestAvgMaxSubtree, plot the result
}

