% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_util.R
\name{plotScndFrstOrdrMinDepth}
\alias{plotScndFrstOrdrMinDepth}
\title{Plot Second Order vs First Order, minimal maximal subtree depth}
\usage{
plotScndFrstOrdrMinDepth(eval_data)
}
\arguments{
\item{eval_data, }{the output of forestAvgMaxSubtree}
}
\value{
so_vs_fo, a ggplot2 object, which plots first order vs second
order minimal maximal subtree depth
}
\description{
Given eval data from forestAvgMaxSubtree, plot the result
}

