% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_util.R
\name{baseRPlotting}
\alias{baseRPlotting}
\title{Plot Second Order vs First Order minimal maximal subtree depth (base R)}
\usage{
baseRPlotting(eval_data)
}
\arguments{
\item{eval_data, }{the output of forestAvgMaxSubtree}
}
\description{
base r version of plotFirstAndSecondOrderMetric (just an alternative)
}

