% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_traversal_util.R
\name{binForestByDepth}
\alias{binForestByDepth}
\title{Bin Forest by Depth}
\usage{
binForestByDepth(forest)
}
\arguments{
\item{forest}{a ranger object from the ranger package, which was created
with param write.forest set to TRUE. In other words, it must have a 'forest'
property.}
}
\value{
returns a list of two elements. the first is a list which contains a
vector for each independent variable, where each element is that variable's
minimal maximal subtree depth, for one tree. The second
element returned, is a vector of tree lengths 'forest_depths'.
}
\description{
Given a forest object, from the ranger package, this function will bin the
forest into depths. This is a helper function for the 'forestAvgMaxSubtree'
function.
}

