% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_traversal_util.R
\name{forestAvgMaxSubtree}
\alias{forestAvgMaxSubtree}
\title{Forest Averaged Maximal Subtree}
\usage{
forestAvgMaxSubtree(ranger_obj)
}
\arguments{
\item{ranger_obj}{a ranger object from the ranger package, which was created
with param
write.forest set to TRUE. In other words, it must have a 'forest' property.}
}
\value{
a data.frame with two columns: minimal maximal subtree depth (first
and second order)
}
\description{
Given an object of class ranger, from the Ranger package (write.forest must
be set to TRUE), this function will traverse the trees and calculate the
minimal maximal subtree metric (first and second order)
}

