% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_util.R
\name{plotAMDMSvsNumSplits}
\alias{plotAMDMSvsNumSplits}
\title{Plot metric vs number of splits per variable.}
\usage{
plotAMDMSvsNumSplits(eval_data, add_text_labels = FALSE)
}
\arguments{
\item{eval_data}{The output of calculateAMDMS.}

\item{add_text_labels}{An optional parameter to turn on text labels next
to the feature's dot on the plot.}
}
\value{
plot object
}
\description{
Plots the metric against the number of splits in the forest. Features with
many splits will be weighed unfairly by the metric.
}

